shift config parsing into shared
