+++ crates/orderbook/src/arguments.rs
@@ -1,12 +1,10 @@
-use anyhow::{anyhow, Context, Result};
-use model::app_id::AppId;
 use primitive_types::{H160, U256};
 use reqwest::Url;
 use shared::{
-    arguments::display_option, bad_token::token_owner_finder, fee_subsidy::cow_token::SubsidyTiers,
-    http_client, price_estimation::PriceEstimatorType, rate_limiter::RateLimitingStrategy,
+    arguments::display_option, bad_token::token_owner_finder, http_client,
+    price_estimation::PriceEstimatorType, rate_limiter::RateLimitingStrategy,
 };
-use std::{collections::HashMap, net::SocketAddr, num::NonZeroUsize, time::Duration};
+use std::{net::SocketAddr, num::NonZeroUsize, time::Duration};
 
 #[derive(clap::Parser)]
 pub struct Arguments {
@@ -49,24 +47,6 @@ pub struct Arguments {
         parse(try_from_str = shared::arguments::duration_from_seconds),
     )]
     pub max_order_validity_period: Duration,
-
-    /// The time period an EIP1271-quote request is valid.
-    #[clap(
-        long,
-        env,
-        default_value = "600",
-        parse(try_from_str = shared::arguments::duration_from_seconds),
-    )]
-    pub eip1271_onchain_quote_validity_seconds: Duration,
-
-    /// The time period an PRESIGN-quote request is valid.
-    #[clap(
-        long,
-        env,
-        default_value = "600",
-        parse(try_from_str = shared::arguments::duration_from_seconds),
-    )]
-    pub presign_onchain_quote_validity_seconds: Duration,
 
     /// The amount of time in seconds a classification of a token into good or bad is valid for.
     #[clap(
@@ -114,51 +94,6 @@ pub struct Arguments {
     #[clap(long, default_value = "24")]
     pub solvable_orders_max_update_age_blocks: u64,
 
-    /// A flat fee discount denominated in the network's native token (i.e. Ether for Mainnet).
-    ///
-    /// Note that flat fee discounts are applied BEFORE any multiplicative factors from either
-    /// `--fee-factor` or `--partner-additional-fee-factors` configuration.
-    #[clap(long, env, default_value = "0")]
-    pub fee_discount: f64,
-
-    /// The minimum value for the discounted fee in the network's native token (i.e. Ether for
-    /// Mainnet).
-    ///
-    /// Note that this minimum is applied BEFORE any multiplicative factors from either
-    /// `--fee-factor` or `--partner-additional-fee-factors` configuration.
-    #[clap(long, env, default_value = "0")]
-    pub min_discounted_fee: f64,
-
-    /// Gas Fee Factor: 1.0 means cost is forwarded to users alteration, 0.9 means there is a 10%
-    /// subsidy, 1.1 means users pay 10% in fees than what we estimate we pay for gas.
-    #[clap(long, env, default_value = "1", parse(try_from_str = shared::arguments::parse_unbounded_factor))]
-    pub fee_factor: f64,
-
-    /// Used to specify additional fee subsidy factor based on app_ids contained in orders.
-    /// Should take the form of a json string as shown in the following example:
-    ///
-    /// '0x0000000000000000000000000000000000000000000000000000000000000000:0.5,$PROJECT_APP_ID:0.7'
-    ///
-    /// Furthermore, a value of
-    /// - 1 means no subsidy and is the default for all app_data not contained in this list.
-    /// - 0.5 means that this project pays only 50% of the estimated fees.
-    #[clap(
-        long,
-        env,
-        default_value = "",
-        parse(try_from_str = parse_partner_fee_factor),
-    )]
-    pub partner_additional_fee_factors: HashMap<AppId, f64>,
-
-    /// Used to configure how much of the regular fee a user should pay based on their
-    /// COW + VCOW balance in base units on the current network.
-    ///
-    /// The expected format is "10:0.75,150:0.5" for 2 subsidy tiers.
-    /// A balance of [10,150) COW will cause you to pay 75% of the regular fee and a balance of
-    /// [150, inf) COW will cause you to pay 50% of the regular fee.
-    #[clap(long, env)]
-    pub cow_fee_factors: Option<SubsidyTiers>,
-
     /// The API endpoint to call the mip v2 solver for price estimation
     #[clap(long, env)]
     pub quasimodo_solver_url: Option<Url>,
@@ -200,15 +135,6 @@ pub struct Arguments {
     )]
     pub amount_to_estimate_prices_with: Option<U256>,
 
-    #[clap(
-        long,
-        env,
-        default_value = "Baseline",
-        arg_enum,
-        use_value_delimiter = true
-    )]
-    pub price_estimators: Vec<PriceEstimatorType>,
-
     /// How many successful price estimates for each order will cause a fast price estimation to
     /// return its result early.
     /// The bigger the value the more the fast price estimation performs like the optimal price
@@ -226,15 +152,6 @@ pub struct Arguments {
     /// max_back_off: f64 in seconds
     #[clap(long, env, verbatim_doc_comment)]
     pub price_estimation_rate_limiter: Option<RateLimitingStrategy>,
-
-    /// The configured addresses whose orders should be considered liquidity and
-    /// not regular user orders.
-    ///
-    /// These orders have special semantics such as not being considered in the
-    /// settlements objective funtion, not receiving any surplus, and being
-    /// allowed to place partially fillable orders.
-    #[clap(long, env, use_value_delimiter = true)]
-    pub liquidity_order_owners: Vec<H160>,
 
     /// The API endpoint for the Balancer SOR API for solving.
     #[clap(long, env)]
@@ -261,16 +178,6 @@ impl std::fmt::Display for Arguments {
         )?;
         writeln!(
             f,
-            "eip1271_onchain_quote_validity_second: {:?}",
-            self.eip1271_onchain_quote_validity_seconds
-        )?;
-        writeln!(
-            f,
-            "presign_onchain_quote_validity_second: {:?}",
-            self.presign_onchain_quote_validity_seconds
-        )?;
-        writeln!(
-            f,
             "token_quality_cache_expiry: {:?}",
             self.token_quality_cache_expiry
         )?;
@@ -290,15 +197,6 @@ impl std::fmt::Display for Arguments {
             "solvable_orders_max_update_age_blocks: {}",
             self.solvable_orders_max_update_age_blocks,
         )?;
-        writeln!(f, "fee_discount: {}", self.fee_discount)?;
-        writeln!(f, "min_discounted_fee: {}", self.min_discounted_fee)?;
-        writeln!(f, "fee_factor: {}", self.fee_factor)?;
-        writeln!(
-            f,
-            "partner_additional_fee_factors: {:?}",
-            self.partner_additional_fee_factors
-        )?;
-        writeln!(f, "cow_fee_factors: {:?}", self.cow_fee_factors)?;
         display_option(f, "quasimodo_solver_url", &self.quasimodo_solver_url)?;
         display_option(f, "yearn_solver_url", &self.yearn_solver_url)?;
         writeln!(
@@ -321,7 +219,6 @@ impl std::fmt::Display for Arguments {
             "amount_to_estimate_prices_with",
             &self.amount_to_estimate_prices_with,
         )?;
-        writeln!(f, "price_estimators: {:?}", self.price_estimators)?;
         writeln!(
             f,
             "fast_price_estimation_results_required: {}",
@@ -332,42 +229,9 @@ impl std::fmt::Display for Arguments {
             "price_estimation_rate_limites",
             &self.price_estimation_rate_limiter,
         )?;
-        writeln!(
-            f,
-            "liquidity_order_owners: {:?}",
-            self.liquidity_order_owners
-        )?;
         display_option(f, "balancer_sor_url", &self.balancer_sor_url)?;
         Ok(())
     }
-}
-
-/// Parses a comma separated list of colon separated values representing fee factors for AppIds.
-fn parse_partner_fee_factor(s: &str) -> Result<HashMap<AppId, f64>> {
-    let mut res = HashMap::default();
-    if s.is_empty() {
-        return Ok(res);
-    }
-    for pair_str in s.split(',') {
-        let mut split = pair_str.trim().split(':');
-        let key = split
-            .next()
-            .ok_or_else(|| anyhow!("missing AppId"))?
-            .trim()
-            .parse()
-            .context("failed to parse address")?;
-        let value = split
-            .next()
-            .ok_or_else(|| anyhow!("missing value"))?
-            .trim()
-            .parse::<f64>()
-            .context("failed to parse fee factor")?;
-        if split.next().is_some() {
-            return Err(anyhow!("Invalid pair lengths"));
-        }
-        res.insert(key, value);
-    }
-    Ok(res)
 }
 
 #[cfg(test)]
